Criando testes de serviço com RESTassured.
